fix: correct Int8 overflow in hour*60 session-minute arithmetic
Polars' dt.hour() returns Int8, so hour*60 silently overflowed
(9*60=540 wraps in 8-bit signed arithmetic) wherever session-minute-
of-day was computed directly from it. This corrupted 5-minute bucket
assignment in aggregate_1m_to_5m (the known CI aggregation failure)
and the same pattern in time_features, microstructure intrabar
features, and the 05_labels pipeline session filter. Cast to Int32
before the multiply in all four call sites.

Also extends aggregate_1m_to_5m's output with session_date and
source_1m_count, which src/features/catalog.py already expects as
non-feature columns but the aggregator was silently dropping.

Also gitignore fyers_auth.py and move the live token file into
secrets/ (already gitignored) after removing hardcoded FYERS
credentials from fyers_auth.py in favor of required env vars.

+++ src/features/microstructure.py
@@ -10,7 +10,8 @@ def add_1m_inside_5m_features(
     one = df_1m.sort("timestamp").with_columns(
         pl.col("timestamp").dt.date().alias("_date"),
         (
-            pl.col("timestamp").dt.hour() * 60 + pl.col("timestamp").dt.minute()
+            pl.col("timestamp").dt.hour().cast(pl.Int32) * 60
+            + pl.col("timestamp").dt.minute().cast(pl.Int32)
         ).alias("_minute"),
         (pl.col("close") > pl.col("open")).cast(pl.Int8).alias("_up"),
         (pl.col("close") < pl.col("open")).cast(pl.Int8).alias("_down"),
@@ -19,7 +20,8 @@ def add_1m_inside_5m_features(
     five = df_5m.sort("timestamp").with_columns(
         pl.col("timestamp").dt.date().alias("_date"),
         (
-            pl.col("timestamp").dt.hour() * 60 + pl.col("timestamp").dt.minute()
+            pl.col("timestamp").dt.hour().cast(pl.Int32) * 60
+            + pl.col("timestamp").dt.minute().cast(pl.Int32)
         ).alias("_close_minute"),
     )
     summary = (
@@ -61,24 +63,12 @@ def add_1m_inside_5m_features(
             pl.len().alias("f_1m_count"),
         )
         .with_columns(
-            (
-                pl.col("_last_1m_close") - pl.col("_first_1m_open")
-            ).alias("f_1m_path_net_move"),
-            (
-                pl.col("f_1m_path_high") - pl.col("f_1m_path_low")
-            ).alias("f_1m_path_range"),
-            (
-                pl.col("f_1m_path_high") - pl.col("_last_1m_close")
-            ).alias("f_1m_path_high_rejection"),
-            (
-                pl.col("_last_1m_close") - pl.col("f_1m_path_low")
-            ).alias("f_1m_path_low_rebound"),
-            (
-                pl.col("f_1m_up_count") / (pl.col("f_1m_count") + 1e-9)
-            ).alias("f_1m_up_ratio"),
+            (pl.col("_last_1m_close") - pl.col("_first_1m_open")).alias("f_1m_path_net_move"),
+            (pl.col("f_1m_path_high") - pl.col("f_1m_path_low")).alias("f_1m_path_range"),
+            (pl.col("f_1m_path_high") - pl.col("_last_1m_close")).alias("f_1m_path_high_rejection"),
+            (pl.col("_last_1m_close") - pl.col("f_1m_path_low")).alias("f_1m_path_low_rebound"),
+            (pl.col("f_1m_up_count") / (pl.col("f_1m_count") + 1e-9)).alias("f_1m_up_ratio"),
         )
         .drop(["_first_1m_open", "_last_1m_close"])
     )
-    return five.join(summary, on="timestamp", how="left").drop(
-        ["_date", "_close_minute"]
-    )
+    return five.join(summary, on="timestamp", how="left").drop(["_date", "_close_minute"])
